requirements.txt
streamlit>=1.28.0
python-pptx==0.6.23
lxml==4.9.3
Pillow==10.0.0
